Moved more tests into specific folders
