+++ Tests/Sudoku/GameState/GameStateMemoryTests.cs
@@ -2,7 +2,7 @@
 using XenobiaSoft.Sudoku;
 using XenobiaSoft.Sudoku.GameState;
 
-namespace UnitTests.Sudoku;
+namespace UnitTests.Sudoku.GameState;
 
 public class GameStateMemoryTests : BaseTestByAbstraction<GameStateMemory, IGameStateMemory>
 {
